Update model buyer ui [WIP]
