+++ model_buyer/ui/src/requirements-form/model.js
@@ -1,0 +1,21 @@
+const model = {
+    model_type: '',
+    data_requirements: {
+        features: {
+            list: [],
+            range: [],
+            pre_processing: [
+                {
+                    method: '',
+                    parameters: ''
+                }
+            ],
+            desc: {}
+        },
+        target: {
+            range: [],
+            desc: []
+        }
+    }
+}
+export default model;
