+++ model_buyer/ui/src/requirements-form/styles.js
@@ -1,0 +1,40 @@
+const styles = theme => ({
+    container: {
+        display: 'flex',
+        flexWrap: 'wrap',
+    },
+    textField: {
+        marginLeft: theme.spacing.unit,
+        marginRight: theme.spacing.unit,
+        width: 500,
+    },
+    dense: {
+        marginTop: 19,
+    },
+    menu: {
+        width: 200,
+    },
+    button: {
+        margin: theme.spacing.unit,
+    },
+    input: {
+        display: 'none',
+    },
+
+    formControl: {
+        margin: theme.spacing.unit,
+        minWidth: 120,
+    },
+    root: {
+        backgroundColor: theme.palette.background.paper,
+    },
+    listSection: {
+        backgroundColor: 'inherit',
+    },
+    ul: {
+        backgroundColor: 'inherit',
+        padding: 0,
+    },
+});
+
+export default styles;
